+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * Drawer `plot.Lines` for plotting table observer data, with a line series per column, and a common X column (#22)
 * Drawer `plot.Scatter` for plotting table observer data as scatter plots. Supports multiple observers and multiple series per observer (#25)
 * Drawer `plot.Bars` for plotting row observer data as bar chart (#27)
+* Drawer `plot.Contour` for plotting grid data as contours (#32)
+* Drawer `plot.HeatMap` for plotting grid data as heat maps (#32)
 * Plot title, axes labels and axes limits can be configured for plots (optional) (#30)
 * Optional selection of columns in bar and time series plots (#30)